added media support and fixed markdown errors

+++ helper.py
@@ -2,6 +2,7 @@ import os, aiofiles, re, aiohttp, string
 import discord,random, json, traceback
 from telebot.util import escape
 from telebot import types
+from typing import Union
 from moviepy.editor import VideoFileClip
 import function as func 
 BOT_TOKEN = func.tokens.dctoken
@@ -43,9 +44,7 @@ async def sendAttachments(message, tgbot, TELEGRAM_CHAT_ID, reply_params=None):
                 media_group.append(types.InputMediaDocument(attachment.url))
         
         # Construct the message header with author and channel information
-        header = f"{message.author.name} | #{message.channel.name}\n\n`{message.id}` | `{message.channel.id}`"
-        # Escape markdown characters in the header
-        content = escapeMD(header)
+        content = f"{escapeMD(message.author.name)} \\| \\#{escapeMD(message.channel.name)}\n\n`{message.id}` \\| `{message.channel.id}`"
         # Send the header as a Markdown-formatted message
         await tgbot.send_message(TELEGRAM_CHAT_ID, content, parse_mode='markdownv2')
         # Send the media group to the specified Telegram chat
@@ -63,8 +62,7 @@ async def sendEmoji(tgbot, message, items, reply_params=None):
         msgcontent, header, TELEGRAM_CHAT_ID = items
         # Check if message has single emoji
         if len(msgcontent) == 1:
-            header = header + f"\n`{message.id}` | `{message.channel.id}`"
-            caption = escapeMD(header)
+            caption = header + f"\n`{message.id}` | `{message.channel.id}`"
             # Determine if the emoji is a PNG or GIF and send the emoji accordingly
             if ".png" in msgcontent[0]:
                 msg = await tgbot.send_photo(TELEGRAM_CHAT_ID, msgcontent[0], caption=caption, parse_mode = "markdownv2", reply_parameters = reply_params)
@@ -97,8 +95,7 @@ async def sendAnimation(session, tgbot, message, items, reply_params = None):
         # If the animation is from Tenor, get the direct URL
         if "tenor" in msgcontent:
             msgcontent = await get_direct_gif_url(session, msgcontent.strip())
-        header = header + f"\n`{message.id}` | `{message.channel.id}`"
-        caption = escapeMD(header)
+        caption = header + f"\n`{message.id}` | `{message.channel.id}`"
         
         # Send the animation with the provided caption
         msg = await tgbot.send_animation(TELEGRAM_CHAT_ID, msgcontent, caption=caption, parse_mode="markdownv2", reply_parameters = reply_params)
@@ -133,7 +130,7 @@ async def downloadFile(session, url):
             content_type = response.headers['Content-Type']
             file_extension = getExtension(content_type)
 
-            if file_extension:
+            if file_extension!='':
                 filename = generate_random_filename(extension=file_extension)
                 save_path = os.path.join("telegramdownloads", filename)
                 content = await response.read()
@@ -141,18 +138,29 @@ async def downloadFile(session, url):
                 return save_path
     return None
 
-async def getSavepath(file_id, animation: bool = False):
+async def getSavepath(session, file_id, tgbot, animation: bool = False):
     # Get the file information using the file ID
     file_info = await tgbot.get_file(file_id)
     filepath = file_info.file_path
+    file_extension = filepath.split('.')[-1]
+    downloaded_file = await tgbot.download_file(filepath)
+    #mime = magic.Magic(mime=True)
+    #file_mime_type = mime.from_buffer(downloaded_file)
+    filename = generate_random_filename(extension=file_extension)
+    save_path = os.path.join("telegramdownloads", filename)
+    
 
     # Construct the URL to download the file
-    file_url = f'https://api.telegram.org/file/bot{func.tokens.tgtoken}/{filepath}'
-    save_path = await downloadFile(session, file_url)
+    #file_url = f'https://api.telegram.org/file/bot{func.tokens.tgtoken}/{filepath}'
+    #save_path = await downloadFile(session, file_url)
+    await saveFile(save_path, downloaded_file)
+    
 
-    if save_path
+    if save_path:
         if animation:
             # Convert video to GIF
+            gifname = generate_random_filename(extension=".gif")
+            gifpath = os.path.join("telegramdownloads", gifname)
             videoClip = VideoFileClip(save_path)
             videoClip.write_gif(gifpath)
             return save_path, gifpath
@@ -166,27 +174,40 @@ async def getTGMedia(session, tgbot, message, animation: bool = False):
             os.makedirs("telegramdownloads")
 
         if message.animation:
-            videopath, filepath = await getSavepath(message.animation.file_id)
-            await delete_file(videopath)
+            if message.animation.file_size > 1048576:
+                filepath = await getSavepath(session, message.animation.file_id, tgbot)
+            else:
+                videopath, filepath = await getSavepath(session, message.animation.file_id, tgbot, True)
+                await delete_file(videopath)
             return filepath
 
+        elif message.audio:
+            filepath = await getSavepath(session, message.audio.file_id, tgbot)
+            filename = message.audio.file_name
+            new_file_path = os.path.join("telegramdownloads", filename)
+            os.rename(filepath, new_file_path)
+            return new_file_path
+
         elif message.video:
-            filepath = await getSavepath(message.video.file_id)
+            filepath = await getSavepath(session, message.video.file_id, tgbot)
             return filepath
 
         elif message.voice:
-            filepath = await getSavepath(message.voice.file_id)
-            return filepath
+            filepath = await getSavepath(session, message.voice.file_id, tgbot)
+            filename = message.voice.file_name
+            new_file_path = os.path.join("telegramdownloads", filename)
+            os.rename(filepath, new_file_path)
+            return new_file_path
         
         elif message.photo:
             filepaths = []
-            for photo in message.photo
-                filepath = await getSavepath(photo.file_id)
+            for photo in message.photo:
+                filepath = await getSavepath(session, photo.file_id, tgbot)
                 filepaths.append(filepath)
             return filepaths
 
         elif message.document:
-            filepath = await getSavepath(message.document.file_id)
+            filepath = await getSavepath(session, message.document.file_id, tgbot)
             return filepath
         
         else:
@@ -197,14 +218,13 @@ async def getTGMedia(session, tgbot, message, animation: bool = False):
         traceback.print_exc()
         return None, None
     
-    
-
 def escapeMD(text):
+    escape_chars = ['_', '*', '[', ']', '(', ')', '~', '`', '>', '#', '+', '-', '=', '|', '{', '}', '. ', '!']
     markdown_symbols = ["-", "|", "#", "+", "."]
     escaped_text = ""
 
     for char in text:
-        if char in markdown_symbols:
+        if char in escape_chars:
             escaped_text += "\\" + char
         else:
             escaped_text += char
@@ -366,38 +386,37 @@ async def send_media(session, channel_id, author, file_path: Union[list, str], r
 
     # Create the payload JSON as a separate field
     payload_json = {"content": ""}
-
-    if isinstance(file_path, str):
-        file = open(file_path, 'rb')
-        file_name = os.path.basename(file_path)
-        file_extension = "."+file_path.split(".",-1)[1]
-        if file_extension in image_extensions:
-            payload_json["embeds"] = [{
-                "image": {
-                    "url": f"attachment://{file_name}"
-                },
-                "color": func.settings.embed_color,
-                "author": {
-                    "name": author
-                }
-            }]
-        data.add_field('file', file, filename=file_name)
-
-    else:
-        for i, file_path in enumerate(file_path):
-            with open(file_path, 'rb') as file:
-                file_name = os.path.basename(file_path)
-                file_extension = "."+file_name.split(".",-1)[1]
-                data.add_field(f'file{i}', file, filename=file_name)
-    
     if reply_messageid:
         payload_json["message_reference"] = {
             "message_id": reply_messageid
         }
+    
+    files = []
+
+    if isinstance(file_path, str):
+        file = open(file_path, 'rb')
+        files.append(file)
+        file_name = os.path.basename(file_path)
+        file_extension = "."+file_name.rsplit(".",1)[1]
+        if file_extension in image_extensions:
+            payload_json["content"] = f"by {author}"
+        data.add_field('files[0]', file, filename=file_name)
+
+    elif isinstance(file_path, list):
+        payload_json["content"] = f"by {author}"
+        index = 0
+        for i in range(2, len(file_path), 3):
+            file = open(file_path[i], 'rb')
+            file_name = os.path.basename(file_path[i])
+            files.append(file)
+            data.add_field(f'files{index}', file, filename=file_name)
+            index+=1
+            
     data.add_field('payload_json', json.dumps(payload_json))
 
     async with session.post(url, data=data, headers=headers) as response:
-        file.close()
+        for file in files:
+            file.close()
         if response.status == 200:
             response_json = await response.json()
             m_id = response_json.get('id')
